pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "payslip2budget"
version = "0.1.0"
description = "Convert payslip PDF to budget transactions (YNAB, Mint, EveryDollar, Monarch)"
readme = "README.md"
requires-python = ">=3.7"
authors = [
    {name = "Coty Sutherland"},
]
license = "MIT"
dependencies = [
  "PyPDF2",
  "pdfplumber",
  "requests",
  "google-api-python-client",
  "google-auth-oauthlib"
]

[project.optional-dependencies]
test = ["pytest"]

[project.scripts]
payslip2budget = "cli:main"

[tool.setuptools]
packages = ["payslip2budget"]